chore: better test renames
