--- apps/service-scrape/lib-db_service_scrape/src/log.ts
@@ -1,3 +1,0 @@
-import pino from 'pino'
-
-export const LOG = pino()
