+++ apps/service-scrape/lib-db_service_scrape/test/connection.test.ts
@@ -1,13 +1,13 @@
-import { describe, expect, test } from 'vitest'
-import { getKyselyPostgresDb } from '../src/kysely'
+import { describe, expect, it } from 'vitest'
+import { getKyselyPostgresDb } from '../src/connection'
 
 describe('postgres connection test', () => {
-    test('invalid uri should return null', async () => {
+    it('should return null with invalid uri', async () => {
         const db = await getKyselyPostgresDb('invalid uri')
         expect(db).toBeNull()
     })
 
-    test('inactive uri should return null', async () => {
+    it('should return null with inactive uri', async () => {
         const db = await getKyselyPostgresDb('postgresql://user:password@localhost:80/invalid_db')
         expect(db).toBeNull()
     })
